# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SummitMate API
  description: 嘉明湖登山行程助手 - 後端 API
  version: 0.1.0
  contact:
    name: SummitMate
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /trips/{tripId}/messages:
    get:
      summary: Get Trip Messages
      description: 取得行程留言板內容
      operationId: listTripMessages
      tags:
        - trip-messages
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: 頁碼 (1-indexed)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: 成功取得留言列表與分頁資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePaginationResponse"
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該行程
    post:
      summary: Add Trip Message
      description: 新增行程留言
      operationId: addTripMessage
      tags:
        - trip-messages
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "201":
          description: 成功新增留言
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該行程
  /trips/{tripId}/messages/{messageId}:
    put:
      summary: Update Trip Message
      description: 更新行程留言
      operationId: updateTripMessage
      tags:
        - trip-messages
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
        "403":
          description: 無權限編輯此留言
    delete:
      summary: Delete Trip Message
      description: 刪除行程留言
      operationId: deleteTripMessage
      tags:
        - trip-messages
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 刪除成功
        "401":
          description: 未登入
        "403":
          description: 無權限刪除此留言
  /trips/{tripId}/polls:
    get:
      summary: List Trip Polls
      description: 取得行程的所有投票
      operationId: listTripPolls
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: 頁碼 (1-indexed)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功取得投票清單與分頁資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollPaginationResponse"
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該行程
    post:
      summary: Create Trip Poll
      description: 建立行程新投票
      operationId: createTripPoll
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollRequest"
      responses:
        "201":
          description: 成功建立投票
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
        "403":
          description: 無權限存取
  /trips/{tripId}/polls/{pollId}:
    get:
      summary: Get Trip Poll
      description: 取得單一投票詳情與選項
      operationId: getTripPoll
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pollId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 成功取得投票詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該投票
    delete:
      summary: Delete Trip Poll
      description: 刪除投票
      operationId: deleteTripPoll
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pollId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 刪除成功
        "401":
          description: 未登入
        "403":
          description: 無權限刪除
  /trips/{tripId}/polls/{pollId}/options:
    post:
      summary: Add Poll Option
      description: 新增投票選項
      operationId: addPollOption
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pollId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollOptionRequest"
      responses:
        "201":
          description: 成功新增選項
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollOption"
        "400":
          description: 參數/權限錯誤 (例如不允許新增選項)
        "401":
          description: 未登入
        "404":
          description: 找不到該投票
  /trips/{tripId}/polls/{pollId}/options/{optionId}/vote:
    post:
      summary: Vote for Poll Option
      description: 投給特定選項 (若已投票可能會取消或修改，根據業務邏輯決定)
      operationId: votePollOption
      tags:
        - trip-polls
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pollId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: optionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 投票成功 (回傳更新後的投票狀態)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: 狀態錯誤 (已結束投票、不允許多選等)
        "401":
          description: 未登入
        "404":
          description: 找不到該投票或選項
  /favorites:
    get:
      summary: List Favorites
      description: 取得使用者的所有收藏清單
      operationId: listFavorites
      tags:
        - favorites
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 頁碼 (1-indexed)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: 成功取得收藏清單與分頁資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritePaginationResponse"
        "401":
          description: 未登入
    post:
      summary: Add to Favorites
      description: 將項目加入收藏
      operationId: addFavorite
      tags:
        - favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequest"
      responses:
        "201":
          description: 成功加入收藏
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
        "409":
          description: 此項目已經在收藏中
  /favorites/batch:
    post:
      summary: Batch Update Favorites
      description: 批量更新最愛狀態 (離線同步用)
      operationId: batchUpdateFavorites
      tags:
        - favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/BatchFavoriteRequest"
      responses:
        "200":
          description: 批量更新成功
        "400":
          description: 無效的請求
        "401":
          description: 未登入
  /favorites/{targetId}:
    delete:
      summary: Remove from Favorites
      description: 將項目從收藏中移除
      operationId: removeFavorite
      tags:
        - favorites
      security:
        - bearerAuth: []
      parameters:
        - name: targetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 成功移除
        "401":
          description: 未登入
        "404":
          description: 找不到該收藏紀錄
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      required:
        - id
        - trip_id
        - user_id
        - display_name
        - category
        - content
        - timestamp
        - created_at
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar:
          type: string
        category:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MessageRequest:
      type: object
      required:
        - content
      properties:
        parent_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          maxLength: 50
        content:
          type: string
          minLength: 1
    PollOption:
      type: object
      required:
        - id
        - poll_id
        - text
        - vote_count
        - created_at
        - created_by
        - updated_at
        - updated_by
      properties:
        id:
          type: string
          format: uuid
        poll_id:
          type: string
          format: uuid
        text:
          type: string
        vote_count:
          type: integer
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          format: uuid
        voters:
          type: array
          items:
            type: string
            format: uuid
    PollOptionRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
    Poll:
      type: object
      required:
        - id
        - trip_id
        - title
        - is_allow_add_option
        - max_option_limit
        - allow_multiple_votes
        - result_display_type
        - status
        - created_at
        - created_by
        - updated_at
        - updated_by
        - description
        - options
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        is_allow_add_option:
          type: boolean
        max_option_limit:
          type: integer
        allow_multiple_votes:
          type: boolean
        result_display_type:
          type: string
        status:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          format: uuid
    PollRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        is_allow_add_option:
          type: boolean
          default: false
        max_option_limit:
          type: integer
          default: 20
        allow_multiple_votes:
          type: boolean
          default: false
        result_display_type:
          type: string
          default: "realtime"
    Favorite:
      type: object
      required:
        - id
        - target_id
        - type
        - created_at
        - created_by
        - updated_at
        - updated_by
      properties:
        id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
        type:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          format: uuid
    FavoriteRequest:
      type: object
      required:
        - target_id
        - type
      properties:
        target_id:
          type: string
          format: uuid
        type:
          type: string
          maxLength: 30
    BatchFavoriteRequest:
      type: object
      required:
        - target_id
        - type
        - is_favorite
      properties:
        target_id:
          type: string
          format: uuid
        type:
          type: string
          maxLength: 30
        is_favorite:
          type: boolean
...
    PaginationMetadata:
      type: object
      required: [has_more, page, limit, total]
      properties:
        has_more:
          type: boolean
          description: 是否還有下一頁
        page:
          type: integer
          description: 目前頁碼 (1-indexed)
        limit:
          type: integer
          description: 每頁筆數
        total:
          type: integer
          description: 符合條件的總筆數
        next_cursor:
          type: string
          description: 下一頁的指標（cursor-based 相容用，可為 null）
          nullable: true
...
    MessagePaginationResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
    PollPaginationResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Poll"
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
    FavoritePaginationResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Favorite"
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
...